Fixed spacing, brackets, and parens in models
Also finished writing out the get /places route.

Updated NPM start to point to correct file.
