+++ package.json
@@ -2,10 +2,10 @@
   "name": "node-express-gmaps",
   "version": "1.0.0",
   "description": "Node Express Google Places API Practice",
-  "main": "/src/app.js",
+  "main": "./src/app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
+    "start": "nodemon src/app.js"
   },
   "repository": {
     "type": "git",
